+++ flash/__init__.py
@@ -25,7 +25,8 @@ if _TORCH_AVAILABLE:
     from flash.core.data.datasets import FlashDataset, FlashIterableDataset
     from flash.core.data.input_transform import InputTransform
     from flash.core.data.io.output import Output
-    from flash.core.data.process import Postprocess, Preprocess, Serializer
+    from flash.core.data.io.output_transform import OutputTransform
+    from flash.core.data.process import Preprocess, Serializer
     from flash.core.model import Task  # noqa: E402
     from flash.core.trainer import Trainer  # noqa: E402
 
@@ -47,7 +48,7 @@ if _TORCH_AVAILABLE:
         "FlashIterableDataset",
         "InputTransform",
         "Output",
-        "Postprocess",
+        "OutputTransform",
         "Preprocess",
         "Serializer",
         "Task",