+++ flash/core/utilities/types.py
@@ -4,7 +4,8 @@ from torch import nn
 from torchmetrics import Metric
 
 from flash.core.data.io.output import Output
-from flash.core.data.process import Deserializer, Postprocess, Preprocess
+from flash.core.data.io.output_transform import OutputTransform
+from flash.core.data.process import Deserializer, Preprocess
 
 MODEL_TYPE = Optional[nn.Module]
 LOSS_FN_TYPE = Optional[Union[Callable, Mapping, Sequence]]
@@ -15,5 +16,5 @@ LR_SCHEDULER_TYPE = Optional[
 METRICS_TYPE = Union[Metric, Mapping, Sequence, None]
 DESERIALIZER_TYPE = Optional[Union[Deserializer, Mapping[str, Deserializer]]]
 PREPROCESS_TYPE = Optional[Preprocess]
-POSTPROCESS_TYPE = Optional[Postprocess]
+OUTPUT_TRANSFORM_TYPE = Optional[OutputTransform]
 OUTPUT_TYPE = Optional[Output]